# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agrotech API",
    "description": "Api para gerenciamento de plantações e safras",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "OpenAPI Documentation - Development"
    }
  ],
...
  "paths": {
...
    "/api/sensors/{id}": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Fetch one sensor by its id",
        "operationId": "getSensorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SensorResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Sensor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Sensors"
        ],
        "summary": "Update sensor",
        "operationId": "updateSensor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sensor updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SensorResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation error"
          },
          "404": {
            "description": "Sensor not found"
          },
          "409": {
            "description": "Sensor code already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Sensors"
        ],
        "summary": "Delete sensor",
        "operationId": "deleteSensor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sensor deleted"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Sensor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sensor-readings/{id}": {
      "get": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Fetch one sensor reading by its id",
        "operationId": "getReadingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor reading fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Sensor reading not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Update sensor reading",
        "operationId": "updateReading",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorReadingUpdateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sensor reading updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation error"
          },
          "404": {
            "description": "Sensor reading not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Delete sensor reading",
        "operationId": "deleteReading",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sensor reading deleted"
          },
          "400": {
            "description": "Invalid UUID format"
          },
          "404": {
            "description": "Sensor reading not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/sensor-readings": {
      "get": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Fetch all sensor readings with pagination",
        "operationId": "getAllSensorReadings",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerSensorReadingResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor readings fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelSensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Create new sensor reading",
        "operationId": "createSensorReading",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorReadingCreateRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Sensor reading created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Sensor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/sensor-readings/sensor/{sensorId}": {
      "get": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Fetch all readings from a sensor",
        "operationId": "getReadingsBySensor",
        "parameters": [
          {
            "name": "sensorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerSensorReadingResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor readings fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelSensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          },
          "404": {
            "description": "Sensor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/sensor-readings/sensor/{sensorId}/range": {
      "get": {
        "tags": [
          "Sensor Readings"
        ],
        "summary": "Fetch readings from a sensor within a date range",
        "operationId": "getReadingsBySensorAndDateRange",
        "parameters": [
          {
            "name": "sensorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "pagedAssembler",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PagedResourcesAssemblerSensorReadingResponseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensor readings fetched",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelSensorReadingResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters or date range"
          },
          "404": {
            "description": "Sensor not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SensorUpdateRequestDTO": {
        "type": "object",
        "properties": {
          "sensorCode": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3
          },
          "model": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installationDate": {
            "type": "string",
            "format": "date-time"
          },
          "batteryLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "lastMaintenance": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "MAINTENANCE",
              "OFFLINE"
            ]
          }
        }
      },
      "SensorResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sensorCode": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installationDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "MAINTENANCE",
              "OFFLINE"
            ]
          },
          "batteryLevel": {
            "type": "integer",
            "format": "int32"
          },
          "lastMaintenance": {
            "type": "string",
            "format": "date-time"
          },
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "SensorReadingUpdateRequestDTO": {
        "type": "object",
        "properties": {
          "readingTime": {
            "type": "string",
            "format": "date-time"
          },
          "temperature": {
            "type": "number",
            "maximum": 70,
            "minimum": -50
          },
          "humidity": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "soilMoisture": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "windSpeed": {
            "type": "number",
            "minimum": 0
          },
          "windDirection": {
            "type": "string",
            "maxLength": 10,
            "minLength": 0
          },
          "rainfall": {
            "type": "number",
            "minimum": 0
          },
          "soilPh": {
            "type": "number",
            "maximum": 14,
            "minimum": 0
          },
          "lightIntensity": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "SensorReadingResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "readingTime": {
            "type": "string",
            "format": "date-time"
          },
          "temperature": {
            "type": "number"
          },
          "humidity": {
            "type": "number"
          },
          "soilMoisture": {
            "type": "number"
          },
          "windSpeed": {
            "type": "number"
          },
          "windDirection": {
            "type": "string"
          },
          "rainfall": {
            "type": "number"
          },
          "soilPh": {
            "type": "number"
          },
          "lightIntensity": {
            "type": "number"
          },
          "sensorId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "SensorReadingCreateRequestDTO": {
        "type": "object",
        "properties": {
          "readingTime": {
            "type": "string",
            "format": "date-time"
          },
          "temperature": {
            "type": "number",
            "maximum": 70,
            "minimum": -50
          },
          "humidity": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "soilMoisture": {
            "type": "number",
            "maximum": 100,
            "minimum": 0
          },
          "windSpeed": {
            "type": "number",
            "minimum": 0
          },
          "windDirection": {
            "type": "string",
            "maxLength": 10,
            "minLength": 0
          },
          "rainfall": {
            "type": "number",
            "minimum": 0
          },
          "soilPh": {
            "type": "number",
            "maximum": 14,
            "minimum": 0
          },
          "lightIntensity": {
            "type": "number",
            "minimum": 0
          },
          "sensorId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "readingTime",
          "sensorId"
        ]
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedResourcesAssemblerSensorReadingResponseDTO": {
        "type": "object",
        "properties": {
          "forceFirstAndLastRels": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "PagedModelSensorReadingResponseDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "sensorReadingResponseDTOList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SensorReadingResponseDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}